separate tools into video tools and image tools
